(PC-17816) refactor(profile): extract useResetRecreditAmountToShow and useUpdateProfileMutation from api file
