+++ src/features/profile/api/useResetRecreditAmountToShow.ts
@@ -1,0 +1,14 @@
+import { useMutation } from 'react-query'
+
+import { api } from 'api/api'
+import { ResetRecreditAmountToShowMutationOptions } from 'features/profile/types'
+
+export function useResetRecreditAmountToShow({
+  onSuccess,
+  onError,
+}: ResetRecreditAmountToShowMutationOptions) {
+  return useMutation(() => api.postnativev1resetRecreditAmountToShow(), {
+    onSuccess,
+    onError,
+  })
+}
