+++ src/features/profile/api/useUpdateProfileMutation.ts
@@ -4,22 +4,6 @@ import { api } from 'api/api'
 import { UserProfileResponse, UserProfileUpdateRequest } from 'api/gen'
 import { QueryKeys } from 'libs/queryKeys'
 
-export enum CHANGE_EMAIL_ERROR_CODE {
-  TOKEN_EXISTS = 'TOKEN_EXISTS',
-  INVALID_PASSWORD = 'INVALID_PASSWORD',
-  EMAIL_UPDATE_ATTEMPTS_LIMIT = 'EMAIL_UPDATE_ATTEMPTS_LIMIT',
-}
-
-export interface ChangeEmailRequest {
-  email: string
-  password: string
-}
-
-type ResetRecreditAmountToShowMutationOptions = {
-  onSuccess: () => void
-  onError: (error: unknown) => void
-}
-
 export function useUpdateProfileMutation(
   onSuccessCallback: (data: UserProfileResponse) => void,
   onErrorCallback: (error: unknown) => void
@@ -34,15 +18,5 @@ export function useUpdateProfileMutation(
       onSuccessCallback(response)
     },
     onError: onErrorCallback,
-  })
-}
-
-export function useResetRecreditAmountToShow({
-  onSuccess,
-  onError,
-}: ResetRecreditAmountToShowMutationOptions) {
-  return useMutation(() => api.postnativev1resetRecreditAmountToShow(), {
-    onSuccess,
-    onError,
   })
 }
