+++ src/features/profile/pages/ChangeEmail/ChangeEmail.tsx
@@ -11,11 +11,12 @@ import { ApiError } from 'api/apiHelpers'
 import { isLongEnough } from 'features/auth/components/PasswordSecurityRules'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
-import { ChangeEmailRequest, CHANGE_EMAIL_ERROR_CODE } from 'features/profile/api'
 import { AlreadyChangedEmailDisclaimer } from 'features/profile/components/Disclaimers/AlreadyChangedEmailDisclaimer'
 import { ChangeEmailDisclaimer } from 'features/profile/components/Disclaimers/ChangeEmailDisclaimer'
+import { CHANGE_EMAIL_ERROR_CODE } from 'features/profile/enums'
 import { useCheckHasCurrentEmailChange } from 'features/profile/helpers/useCheckHasCurrentEmailChange'
 import { useValidateEmail } from 'features/profile/helpers/useValidateEmail'
+import { ChangeEmailRequest } from 'features/profile/types'
 import { analytics } from 'libs/firebase/analytics'
 import { useSafeState } from 'libs/hooks'
 import { theme } from 'theme'
