+++ src/features/profile/pages/NotificationSettings/NotificationSettings.tsx
@@ -6,7 +6,7 @@ import { checkNotifications, PermissionStatus } from 'react-native-permissions'
 import { NotificationSubscriptions, UserProfileResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { PushNotificationsModal } from 'features/notifications/batchNotifications/components/PushNotificationsModal'
-import { useUpdateProfileMutation } from 'features/profile/api'
+import { useUpdateProfileMutation } from 'features/profile/api/useUpdateProfileMutation'
 import { PageProfileSection } from 'features/profile/components/PageProfileSection/PageProfileSection'
 import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch/SectionWithSwitch'
 import { useAppStateChange } from 'libs/appState'
