+++ src/features/profile/types.ts
@@ -1,0 +1,9 @@
+export interface ChangeEmailRequest {
+  email: string
+  password: string
+}
+
+export type ResetRecreditAmountToShowMutationOptions = {
+  onSuccess: () => void
+  onError: (error: unknown) => void
+}
